==> Code/2_manuscript-and-figures.Rmd <==
---
title: "Comparison of rates of firearm and non-firearm homicide and suicide in black and white non-Hispanic men by US state"
output: 
  word_document:
      reference_docx: reference.docx
---

```{r load-libraries, echo = F, warning = F}
library(tidyverse)  # a collection of packages which we use for data manipulation)
library(ggrepel)  # for formatting the labels on a ggplot with many points
```

```{r load-clean-data, echo = F}
# This loads the dataset created in step 1 into memory. It is called CDC_Males_AllAges_0816_wide
load(file = "../Data/Cleaned-data/Males_AllAges_0816.Rdata")
```

Corinne A Riddell, PhD

Sam Harper, PhD

Magdalena Cerdá, PhD

Jay S Kaufman, PhD

CAR, SH, and JSK: Department of Epidemiology, Biostatistics & Occupational Health, McGill University, 1020 Pine Avenue West, Room 27, Montreal, QC H3A 1A2, Canada.

MC: Violence Prevention Research Program, Department of Emergency Medicine, University of California, Davis, USA.

Address correspondence to: Corinne Riddell, PhD, Department of Epidemiology, Biostatistics & Occupational Health, McGill University, 1020 Pine Avenue West, Room 27, Montreal, QC H3A 1A2, Canada. Email: <corinne.riddell@mail.mcgill.ca>. Phone: +1-438-863-0402

Word Count: 3647

##### PAGE BREAK

#Abstract

**Background:** It is unknown how much the difference in excess deaths due to homicide among black men and suicide among white men vary by US state. 

**Objective:** To compare the rates of firearm and non-firearm homicide and suicide in black and white non-Hispanic men by US state and examine their associations with state prevalence of gun ownership.

**Design:** Surveillance study.

**Setting:** 50 states and the District of Columbia, 2008 to 2016. Cause of death data was abstracted using the CDC WONDER Online Database.

**Participants:** Non-Hispanic black and non-Hispanic white males, all ages.

**Measurements:** Absolute rates and rate differences in firearm and non-firearm homicide and suicide for black and white men.

**Results:** There were 84,113 homicides and 251,772 suicides over the 9-year period. Black-white differences in rates of firearm homicide and firearm suicide varied widely across states. Relative to white men, black men experienced between 9 and 57 additional firearm homicides per 100,000 men per year, with black men in Missouri, Michigan, Illinois, Indiana, and Pennsylvania having more than 40 additional firearm homicides per 100,000 per year. White men experienced between 2 fewer and 16 additional firearm suicides per 100,000 per year, with the largest inequalities observed in Southern and Western states, and the smallest inequalities in the District of Columbia and densely populated Northeastern states.

**Limitations:** Misclassification of homicides and suicides as unintentional injury deaths may have occurred. Survey data on state household gun ownership was collected in 2004 and may have shifted during the past decade.

**Conclusion:** The large variation in the rates of firearm homicide and firearm suicide and their racial inequalities highlight states where policies may be most beneficial at reducing rates of homicide and suicide and their racial inequalities.

**Primary Funding Source:** CAR receives salary support from the Department of Epidemiology, Biostatistics, and Occupational Health. MC is supported by the National Institutes of Health (#1R21 DA041154-01). 

##### PAGE BREAK

#Introduction

```{r CDC-wonder-intro-data, echo = F}
#The introduction to the paper provides the relative risk of homicide and suicide between black and white men. 
#These summary statistics were calculated using data from CDC WONDER that I stored in the Data/Cited-data directory.
#Have a look at the notes inside the data files for the selections made using CDC WONDER to generate these results.

#Read in the homicide data
intro.data.homicide <- read.table("../Data/Cited-data/Homicide_NH_Males_Age-adjusted_2016.txt", 
                                  header = T, nrows = 2, sep = "\t")

#calculate the risk ratio
homicide.stats <- intro.data.homicide %>% select(Race, Age.Adjusted.Rate) %>% 
  spread(key = Race, value = Age.Adjusted.Rate) %>% 
  mutate(RR = `Black or African American`/White) %>% rename(Black = `Black or African American`)

#Read in the *firearm* homicide data
intro.data.homicide.FA <- read.table("../Data/Cited-data/Homicide_Firearms_NH_Males_Age-adjusted_2016.txt", 
                                  header = T, nrows = 2, sep = "\t")

#Calculate and store the portion of homicides that are committed using firearms
prop.homicides.FA <- round(sum(intro.data.homicide.FA$Deaths)/sum(intro.data.homicide$Deaths)*100)

#calculate the risk ratio for firearm homicides
FA.homicide.stats <- intro.data.homicide.FA %>% select(Race, Age.Adjusted.Rate) %>% 
  spread(key = Race, value = Age.Adjusted.Rate) %>% 
  mutate(RR = `Black or African American`/White) %>% rename(Black = `Black or African American`)

#Suicides -- repeat all the above steps.
intro.data.suicide <- read.table("../Data/Cited-data/Suicide_NH_Males_Age-adjusted_2016.txt", 
                                  header = T, nrows = 2, sep = "\t")

suicide.stats <- intro.data.suicide %>% select(Race, Age.Adjusted.Rate) %>% 
  spread(key = Race, value = Age.Adjusted.Rate) %>% 
  mutate(RR = White/`Black or African American`) %>% rename(Black = `Black or African American`)

intro.data.suicide.FA <- read.table("../Data/Cited-data/Suicide_Firearms_NH_Males_Age-adjusted_2016.txt", 
                                  header = T, nrows = 2, sep = "\t")

prop.suicides.FA <- round(sum(intro.data.suicide.FA$Deaths)/sum(intro.data.suicide$Deaths)*100)

FA.suicide.stats <- intro.data.suicide.FA %>% select(Race, Age.Adjusted.Rate) %>% 
  spread(key = Race, value = Age.Adjusted.Rate) %>% 
  mutate(RR = White/`Black or African American`) %>% rename(Black = `Black or African American`)

#We wrote our paper using R markdown and below we reference the saved values inside small code chunks indicated by: `r `. When this file is compiled using `Knit` the code chunks are compiled and replaced by their calculated values.
```

In 2016, non-Hispanic black men were nearly `r round(homicide.stats$RR[1], 1)` times more likely to die from homicide than non-Hispanic white men in the United States, whereas white men were `r round(suicide.stats$RR[1], 1)` times more likely than black men to die from suicide (1). More than `r signif(prop.homicides.FA,1)`% of homicides and `r round(prop.suicides.FA)`% of suicides involved firearms. The relative racial inequalities in violent deaths are especially marked for firearm homicide. Black men are `r round(FA.homicide.stats$RR[1], 0)` times more likely to die of a firearm homicide than white men (1).

Racial inequalities in violent deaths are well-established for the United States overall, but it is unknown whether and how much they vary by state. Large variation in racial inequalities would suggest the presence of explanatory factors that differ across states. These factors could reflect differences in race-specific levels of exposure to socio-economic characteristics, neighborhood residential conditions, medical care, or policy environments that are known to affect violence (2). Measuring variation across states in racial inequalities in homicide and suicide, which are masked when looking only at the population average, can help to identify areas where sub-populations may be experiencing an excess burden of injury, and where efforts could be directed to try and reduce inequalities.  

Across the US, states with the highest levels of household gun ownership have higher rates of firearm suicide and homicide (3). Ecologic studies comparing US states have estimated positive associations between rates of gun ownership and firearm homicide and firearm suicide rates (4,5,6,7,8), and case-control studies have found positive associations between having a gun in the home and an individual’s risk of homicide occurring at home (9), and with the overall risk of suicide (10). When considering the causal pathway to homicide or suicide, two component causes are motivation and opportunity (11). Gun availability or ownership increases opportunity and it is plausible that motivation varies by racial group. For example,  participation in illegal activity is linked to increased homicide risk (12), while more social support and cultural cohesion is linked to decreased suicide risk (13). Thus, the relationship between gun availability and homicide or suicide may be differential by race. This information is currently unknown but could inform whether policies aimed at reducing gun prevalence may have differential effects among blacks and whites.

In this study, we aimed to: (1) Compare the rates of firearm and non-firearm homicide and suicide in black and white non-Hispanic men by US state; (2) Examine the association of state prevalence of gun ownership and firearm homicides and suicides.
  
#Methods

We used the Center for Disease Control and Prevention (CDC) WONDER Online Database to abstract the age-adjusted rates of homicide and suicide (14). Data were abstracted for all states and the District of Columbia between 2008 and 2016 for non-Hispanic white and black males of any age. The tenth revision of the International Classification of Diseases codes was used to identify firearm homicides (X93, X94, X95), non-firearm homicides (X85-X92, X96-X99, and Y01-Y05), firearm suicides (X72, X73, X74), and non-firearm suicides (X60-X71 and X75-X84). Rates were suppressed by the data provider if there were fewer than twenty deaths for the race group in the state between 2008 and 2016. Suppressed data were set to missing in our statistical analysis. These data originated from death certificates provided by states to the National Vital Statistics System, and overseen by the CDC's National Center for Health Statistics. Race and Hispanic ethnicity were recorded by medical examiners. Throughout this paper we refer to this population as men since the vast majority of deaths occurred after childhood.

We computed the average annual fatality rate according to race, intent, and firearm involvement by dividing the number of fatalities by the total population estimate (i.e., the sum of population estimates over 2008 to 2016) and we calculated the absolute inequality (rate difference) in estimates between black and white men. For homicide, we subtract the fatality rate of white men from that of black men. For suicide these rates are reversed, such that the estimated rate differences are mostly positive and interpreted as the number of additional deaths to black (white) men due to homicide (suicide) associated with the heightened fatality rate experienced by black (white) men. When either of the race-specific rates was suppressed, a rate difference was not calculated for the state. To reflect varying population sizes among states and by race, we represent the area of each data marker to be proportional to the race-specific average annualized population size. 

To examine the relationships between household gun ownership, firearm homicide, and firearm suicide by race and state, we used the last available state-level estimates of household firearm ownership from the National Behavioral Risk Factor Surveillance System (BRFSS) from 2004. The BRFSS is an annual telephone survey on the non-institutionalized adult population that asks respondents about their health-related risk behaviors, chronic health conditions, and use of preventive services. In 2001, 2002, and 2004, the BRFSS asked, “Are any firearms kept in or around your home?” More than 293,000 people responded to this question and estimates of household gun ownership prevalence (%) by state were computed previously and ranged between 10.2% in Hawaii and 65.5% in Wyoming (15). Given that individuals may have under-reported gun ownership, and that state gun prevalence may have shifted since 2004, we grouped states into four broad categories of gun prevalence: (i) 10% - 19.9%, (ii) 20% - 34.9%, (iii) 35% - 44.9%, and (iv) 45% - 66%, to provide a reasonable approximation of current levels of ownership. 

We then visually assessed the relationships between state gun prevalence, firearm homicide, and firearm suicide by plotting the rate of firearm homicide as a function of firearm suicide. We color-coded the points by category of gun prevalence, separately for black men and white men. This descriptive exercise illustrates whether states with relatively high rates of firearm homicide also have relatively high rates of firearm suicide, and if this pattern exists for both black men and white men and is associated with state gun ownership. A strong association would suggest that firearm homicide and suicide share risk factors and may be amenable to similar prevention strategies. A weak association would indicate less overlap in risk factors and that distinct preventive approaches may therefore be needed.

Data management and data visualization was performed in R version 3.3.2 (16). No institutional ethical review was required since this study used publicly accessible data (17). A replication data set including the raw data and statistical code to reproduce the manuscript is publicly available on Github: [https://github.com/corinne-riddell/InjuryDisparity/](https://github.com/corinne-riddell/InjuryDisparity/).

**Funding Source**

This research was not formally supported by a research grant, although one of the co-authors (MC) was supported by the National Institutes of Health (#1R21 DA041154-01).

#Results
```{r calculate-number-deaths, echo=F}
#Number of deaths from homicide and suicide for black and white men by type (firearm vs. non-firearm)
num.deaths <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(sum.deaths.homicide.black = sum(Deaths.homicide.Black, na.rm = T), 
            sum.deaths.homicide.white = sum(Deaths.homicide.White, na.rm = T),
            sum.deaths.suicide.black = sum(Deaths.suicide.Black, na.rm = T), 
            sum.deaths.suicide.white = sum(Deaths.suicide.White, na.rm = T))
```

```{r display-suppressed-states, echo=F, eval=F}
#This code lists the suppressed states for the specified conditions. We refer to the sum of these states in the first paragraph of the results.
CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", is.na(adjusted.homicide.Black) == T) %>% select(State)
CDC_Males_AllAges_0816_wide %>% filter(type == "Non-Firearm", is.na(adjusted.homicide.Black) == T) %>% select(State)

CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", is.na(adjusted.homicide.White) == T) %>% select(State)
CDC_Males_AllAges_0816_wide %>% filter(type == "Non-Firearm", is.na(adjusted.homicide.White) == T) %>% select(State)

CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", is.na(adjusted.suicide.Black) == T) %>% select(State)
CDC_Males_AllAges_0816_wide %>% filter(type == "Non-Firearm", is.na(adjusted.suicide.Black) == T) %>% select(State)

CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", is.na(adjusted.suicide.White) == T) %>% select(State) #no states
CDC_Males_AllAges_0816_wide %>% filter(type == "Non-Firearm", is.na(adjusted.suicide.White) == T) %>% select(State) #no states
```

Our analysis included `r formatC(sum(num.deaths$sum.deaths.homicide.black) + sum(num.deaths$sum.deaths.homicide.white), format = "d", big.mark = ",")` homicides and `r formatC(sum(num.deaths$sum.deaths.suicide.black) + sum(num.deaths$sum.deaths.suicide.white), format = "d", big.mark = ",")` suicides from 2008 to 2016. Homicide deaths were suppressed by the National Center for Health Statistics for black men in 16 states and for white men in eight states. Suicide deaths were suppressed in 14 states for black men. There was no suppression of suicide deaths for white men.  

*Comparison of firearm homicide rates*

```{r, echo=F}
#calculate the min and max rate difference (RD) for homicide by type (firearm vs. non-firearm)
stats.RD.homicide <-CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adj.RD.homicide, na.rm = T)), 
            max = round(max(adj.RD.homicide, na.rm = T)))

#calculate the min and max of homicide rates for white men
stats.white.homicide <-CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adjusted.homicide.White, na.rm = T)), 
            max = round(max(adjusted.homicide.White, na.rm = T)))

```

**Figure 1** displays the homicide rates among black men compared to white men, according to state and firearm use. Across the entire US, black men experienced `r CDC_Males_AllAges_0816_wide %>% filter(ST == "US", type == "Firearm") %>% select(adj.RD.homicide) %>% round()` additional homicides per 100,000 per year compared to white men. The black-white difference in firearm homicide rates varied strikingly by state, between `r stats.RD.homicide$min[1]` and `r stats.RD.homicide$max[1]` per 100,000 per year, and was driven by marked variation in firearm homicides to black men across states. States with the smallest black firearm homicide rates saw fewer than 15 deaths per 100,000 per year over the time period, compared to more than 40 deaths per 100,000 per year in states with the highest rates. Firearm homicide rates for white men showed relatively little variation across states, between `r stats.white.homicide$min[1]` and `r stats.white.homicide$max[1]` deaths per 100,000 per year. The states with the highest rates of firearm homicide for black men (e.g., Missouri, Michigan, Illinois, Indiana) also had the largest differences in rates between blacks and whites, signaling states that were most high-risk for blacks but relatively low-risk for whites. DC was also relatively high risk for black men, but not for white men.

*Comparison of firearm homicide rates in black men among states with the largest black populations*

```{r calculate-four-state-statistics, echo = F}
four.states <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST %in% c("FL", "TX", "CA", "GA"), type == "Firearm") %>% 
  select(ST, adj.RD.homicide) %>% 
  arrange(adj.RD.homicide)
```

Another way to make comparisons is to consider the states that are home to the most black residents, to describe whether these states also varied in their rates of firearm homicide and when compared to spatially contiguous states. Nearly 40% of the US black population lives in Georgia, New York, Florida, Texas, and California, and comprise between 6% and 30% of each state's population. Among them, Texas exhibited the smallest black-white difference in firearm homicide (`r four.states %>% filter(ST == "TX") %>% select(adj.RD.homicide) %>% round()` additional deaths to black men per 100,000 per year), which was also substantially lower than the differences estimated for neighboring states like Louisiana (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", ST == "LA") %>% select(adj.RD.homicide) %>% round()`), Oklahoma (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", ST == "OK") %>% select(adj.RD.homicide) %>% round()`), and Arkansas (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Arkansas") %>% select(adj.RD.homicide) %>% round()`). Georgia (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Georgia") %>% select(adj.RD.homicide) %>% round()`) and Florida (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Florida") %>% select(adj.RD.homicide) %>% round()`) had differences similar to nearby South and North Carolina (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "South Carolina") %>% select(adj.RD.homicide) %>% round()` and +`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "North Carolina") %>% select(adj.RD.homicide) %>% round()`, respectively), while having lower rates than Alabama (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Alabama") %>% select(adj.RD.homicide) %>% round()`) and Tennessee (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Tennessee") %>% select(adj.RD.homicide) %>% round()`). New York's difference (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "New York") %>% select(adj.RD.homicide) %>% round()`) was similar to Massachusetts (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Massachusetts") %>% select(adj.RD.homicide) %>% round()`) and Connecticut (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Connecticut") %>% select(adj.RD.homicide) %>% round()`), but was substantially lower than the rates of neighboring New Jersey (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "New Jersey") %>% select(adj.RD.homicide) %>% round()`) and Pennsylvania (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Pennsylvania") %>% select(adj.RD.homicide) %>% round()`). California exhibited the largest black-white difference (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "California") %>% select(adj.RD.homicide) %>% round()`) in firearm homicide among the five states with the largest black populations. California's black-white difference was larger than those estimated for contiguous states like Nevada (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Nevada") %>% select(adj.RD.homicide) %>% round()`) and Oregon (+`r CDC_Males_AllAges_0816_wide %>% filter(type == "Firearm", State == "Oregon") %>% select(adj.RD.homicide) %>% round()`), although these states had relatively small black populations.

*Comparison of non-firearm homicide rates*
```{r, echo=F}
stats.black.homicide <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adjusted.homicide.Black, na.rm = T)), 
            max = round(max(adjusted.homicide.Black, na.rm = T)))

FA.vs.NFA.black.homicide <- CDC_Males_AllAges_0816_wide %>% 
  select(State, adjusted.homicide.Black, type) %>% 
  spread(key = type, value = adjusted.homicide.Black) %>% 
  mutate(RR = Firearm/`Non-Firearm`) %>% arrange(-RR)
```

Non-firearm homicides occur much less frequently for both black men and white men. On average, there were `r CDC_Males_AllAges_0816_wide %>% filter(ST == "US", type == "Non-Firearm") %>% select(adj.RD.homicide) %>% round()` additional non-firearm homicides per 100,000 per year among black men compared to white men. Whereas black men experienced rates that varied between `r stats.black.homicide$min[2]` and `r stats.black.homicide$max[2]` per 100,000 per year by state, white rates were less than `r stats.white.homicide$max[2]` per 100,000 per year. On average, firearm homicide rates were `r FA.vs.NFA.black.homicide %>% filter(State == "United States") %>% select(RR) %>% round()`-fold higher than non-firearm rates for black men. The states with the largest differences between firearm and non-firearm homicide rates were Missouri, Indiana, Illinois, and Louisiana, each of which had firearm homicide rates that were more than twelve times higher than their non-firearm homicide rates.

*Comparison of firearm suicide rates*

```{r, echo=F}
average.RD.suicide <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST == "US", type == "Firearm") %>% 
  select(adj.RD.suicide) %>% round()
```

```{r results_statistics_suicide, echo=F}
stats.RD.suicide <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adj.RD.suicide, na.rm = T)),
            max = round(max(adj.RD.suicide, na.rm = T)))

stats.white.suicide <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adjusted.suicide.White, na.rm = T)),
            max = round(max(adjusted.suicide.White, na.rm = T)))

stats.black.suicide <- CDC_Males_AllAges_0816_wide %>% 
  filter(ST != "US") %>%
  group_by(type) %>% 
  summarise(min = round(min(adjusted.suicide.Black, na.rm = T)),
            max = round(max(adjusted.suicide.Black, na.rm = T)))
```

In contrast to homicide, suicide rates were higher among white men compared to black men, independent of firearms involvement, with the exception of firearm suicide in DC (**Figure 2**). Across the US, there were `r average.RD.suicide*-1` additional firearm suicides among white men per 100,000 per year compared to black men. The white-black difference in firearm suicides varied from `r -(stats.RD.suicide$max[1])` per 100,000 per year in the District of Columbia to `r -(stats.RD.suicide$min[1])` per 100,000 per year in Missouri, although differences could not be estimated for 12 states with fewer than 20 firearm suicides among blacks. The annual rate of firearm suicide varied more markedly for white men than for black men: between `r stats.white.suicide$min[1]` per 100,000 per year in DC and `r stats.white.suicide$max[1]` per 100,000 per year in Wyoming for white men, compared to a range of `r stats.black.suicide$min[1]` per 100,000 per year in Massachusetts to `r stats.black.suicide$max[1]` per 100,000 per year in Alaska among black men. 

Six of the ten states with the largest white-black differences in firearm suicide are located in the South, including Mississippi, Arkansas, Alabama, Texas, South Carolina, and Louisiana, while the remaining four are in the West, including Nevada, Arizona, New Mexico, and Colorado (range: between 12.9 and 15.7 per 100,000 per year). Along with DC, the mid-Atlantic states of Massachusetts, New Jersey, Connecticut, and New York had the smallest white-black differences (4 or less per 100,000 per year) in firearm suicide rates. 

*Comparison of non-firearm suicide rates*
```{r NFA_suicide_stats, echo = F}
NFA_suicide_stats <- CDC_Males_AllAges_0816_wide %>% 
  filter(type == "Non-Firearm") %>% 
  arrange(-adjusted.suicide.White) %>% 
  select(State, adjusted.suicide.White)

NFA_suicide_stats_b <- CDC_Males_AllAges_0816_wide %>% 
  filter(type == "Non-Firearm", !(is.na(adjusted.suicide.Black))) %>%
  arrange(-adjusted.suicide.Black) %>% 
  select(State, adjusted.suicide.Black)
```

The rates of black and white non-firearm suicides and the difference between them varied substantially less than those involving firearms. In the US overall, the non-firearm suicide mortality rate was `r CDC_Males_AllAges_0816_wide %>% filter(ST == "US" & type == "Non-Firearm") %>% select(adjusted.suicide.White) %>% round(1)` per 100,000 per year among white men compared to `r CDC_Males_AllAges_0816_wide %>% filter(ST == "US" & type == "Non-Firearm") %>% select(adjusted.suicide.Black) %>% round(1)` per 100,000 per year among black men. Across the states, `r NFA_suicide_stats[1, 1]` had the highest rate of non-firearm suicides for white men of `r round(NFA_suicide_stats[1,2])` per 100,000 per year, while `r NFA_suicide_stats[nrow(NFA_suicide_stats), 1]` had the lowest rate of `r round(NFA_suicide_stats[nrow(NFA_suicide_stats), 2])` per 100,000 per year, compared to the range estimated for black men from `r round(NFA_suicide_stats_b[1,2])` per 100,000 per year in `r NFA_suicide_stats_b[1,1]` and `r round(NFA_suicide_stats_b[nrow(NFA_suicide_stats_b), 2])` per 100,000 per year in `r NFA_suicide_stats_b[nrow(NFA_suicide_stats_b), 1]`.

```{r calc-spearman-rank-correlation, echo=F}
spearman.white <-CDC_Males_AllAges_0816_wide %>%
  filter(!(ST %in% c("US", "DC", "HI") )) %>%
  summarize(spearmans = cor.test(x = adjusted.suicide.White,
                                 y = adjusted.homicide.White,
                                 method = "spearman")$estimate)
spearman.black <- CDC_Males_AllAges_0816_wide %>%
  filter(!(ST %in% c("US", "DC", "HI") )) %>%
  summarize(spearmans = cor.test(x = adjusted.suicide.Black,
                                 y = adjusted.homicide.Black,
                                 method = "spearman")$estimate)
```

*Relationship between firearm homicide and suicide with state household gun ownership among white men*

Firearm homicide and suicide were strongly associated for white men (**Figure 3**). In terms of the relationship between firearm homicide, suicide, and state gun prevalence, the five states with lower rates of both firearm homicide and suicide among white men (i.e., Massachusetts, New Jersey, Rhode Island, New York, and Connecticut) were also in the lowest category of reported state gun prevalence. Furthermore, many of the states with the highest firearm homicide and suicide rates belonged to the highest category of gun prevalence, while most states located in the middle of the graph belonged to the middle categories of gun prevalence. Notably, several states that belonged to the highest gun prevalence category had relatively low rates of firearm homicide, including Iowa, Nebraska, the Dakotas, Idaho, Montana, and Wyoming. 

*Relationship between firearm homicide and suicide with state household gun ownership among black men*

Among black men, firearm homicide and suicide were modestly associated (**Figure 4**). States with the lowest reported prevalence of guns had the lowest rates of firearm suicide, and below-average rates of firearm homicide. Missouri and several rust-belt states (Illinois, Indiana, Michigan, and Pennsylvania) reported the most firearm homicides among black men and medium levels of gun ownership. Among the Southern states, those with high levels of gun ownership, such as Louisiana, reported higher homicide rates among black men than states with medium levels, such as Texas.

#Discussion

*Key Findings*

Rates of firearm homicide among black men varied substantially across states. For white men, rates were consistently lower and less variable, leading to large racial inequalities in firearm homicide, especially in Missouri, Michigan, Illinois, Indiana, and Pennsylvania. Non-firearm homicides occurred much less frequently than firearm homicides and exhibited little variation across states. However, these homicides still occurred more frequently among black men than white men. White men were at higher risk than black men for firearm suicide, most notably in Mississippi, Nevada, Arizona, New Mexico, Arizona, and Alabama. Northeastern states with reliable data (New York, Connecticut, New Jersey, Massachusetts) and DC exhibited the smallest differences in firearm suicides. These states also had the lowest risk of suicide for both black and white men, implying that they offer the lowest race-specific risks and are the closest to achieving racial equality in firearm suicide. In terms of non-firearm suicide, white men had higher rates than black men, though these rates were less variable than for firearm suicide.

For white men, firearm homicide and suicide shared a strong positive association. This suggests that the characteristics that generate variation across states in firearm homicide and suicide are similar for white men. States with the lowest gun ownership rates had the lowest rates of firearm homicide and firearm suicide among white men, and several states with the highest rates of ownership had the highest rates of firearm homicide and relatively high rates of firearm suicide. 

The relationship between firearm homicide and suicide was modest for black men. In contrast to white men, this moderate relationship suggests that the risk factors for firearm homicide and suicide for black men are more dissimilar, and the substantial variation in firearm homicide among black men suggests the presence of additional explanatory factors for firearm homicide compared to firearm suicide. While state gun ownership was positively associated with both outcomes for white men, some states with medium levels of gun ownership had the highest rates of homicide among black men while others had relatively low rates, implying the presence of important unexamined factors in determining homicide risk for black men.

*Mechanisms*

The spatial distribution of homicides is known to vary, with suicides occurring at the highest rates in the most rural settings and homicides occurring at the highest rates in the most urban settings (18). Particularly with respect to homicide, this urban/rural divide likely reflects predictors present in metropolitan environments such as higher levels of income inequality, female-headed households, and crowding (19). Higher suicide rates in rural areas have been attributed to social and geographic isolation, access to lethal mechanisms, and stresses associated with agricultural work (20). 

While there is ample research on risk factors for homicide and suicide, there is less clarity on the mechanisms that have led to varying racial inequalities across states. Structural disadvantage (i.e., concentrated poverty, joblessness, family disruption) (21), racial residential segregation (22,23,24,25), and participation in drug markets (26,27) have all been implicated in the higher risk of homicide among black men relative to white men. Thus, these risk factors may occur more frequently among black men residing in states with the largest inequalities in firearm homicide.  

Firearm policy has also been implicated in differences in overall rates of homicide and suicide across the US. A cross-sectional ecologic study found that states with more restrictive firearm legislative environments had 6.6 fewer firearm deaths per 100,000 per year than states with the least restrictive environments (28). Analyses that compare changes in states with firearm policy changes to a synthetic control have estimated that policy changes to permit-to-purchase laws are linked with decreases in firearm homicide rates (but not non-firearm homicide) and in suicide rates (29,30). Thus, it is likely that differences in state policy environments have contributed to the levels of homicide and suicide estimates in this study, although it is less clear whether policies are linked to racial inequalities.

State gun prevalence was strongly tied to firearm suicide for both black and white men, and firearms are the most lethal means of attempted suicide (31,32). In addition, states that were the lowest risk for firearm suicide also have implemented the most restrictive gun laws (33). 

Alcohol is another important risk factor for suicide, both in terms of history of alcoholism and overuse at the time of suicide (34,35). Alcohol was found in the systems of 25% of white male suicides tested and 15% of black male suicides in 2005–2006, with 16% of the white men being described as alcohol dependent at the time of suicide compared to 7% of black men (36). Given that white men have higher rates of alcohol abuse than black men (36), alcohol use might be implicated both in the direction of the white-black inequality and its variation across states.

While we were not able to estimate state-level inequalities in age-specific rates of mortality, national figures indicate very different trends for black and white men (37). Suicidal risk increases for white men as they age, and exceeds 30 firearm suicides per 100,000 per year for men 75 and older. For black men, the peak rate of 10 firearm suicides per 100,000 per year occurs at ages 25 to 29 years old (37). The estimated white-black differences in suicide mortality and their variation by state may therefore be reflective of differences across states in the risks of younger black men and older white men to commit suicide by firearm.

*Previous Research*

We searched MEDLINE for literature on racial disparities in homicide or suicide using the search strategy: (“Homicide”[Mesh] OR “Suicide”[Mesh]) AND (health disparit*[Text Word] OR health inequal*[Text Word] OR health inequit*[Text Word] OR social disparit*[Text Word] OR social inequit*[Text Word] OR social inequal*[Text Word]) from database conception to January 19, 2018 with no language restriction. One hundred and two articles were identified, none of which estimated inequalities in homicide and/or suicide mortality by US state. One report did classify states according to their rates of overall homicide mortality (38). This report estimated that overall rates of homicide mortality ranged from 1.1 to 12.8 deaths per 100,000 in 2009, with Southern states generally having the highest rates (38). Given the large differences in homicide mortality by race, our analysis adds to these findings by emphasizing racial inequalities and demonstrating their considerable variation across states.  

*Study Limitations*

This study is subject to several limitations. Some homicide and suicide deaths are difficult to classify. Homicides may be classified as suicides, or vice versa, and either may be misclassified as an unintentional injury. Misclassification of drug poisoning deaths are of particular importance during the opioid epidemic that has led to a sharp increase in fatal overdoses (39). The most recent data on state gun ownership was collected in 2004, four to twelve years prior to the mortality data used in this study. Given that firearm suicide attempts are more than 80% fatal (31,32), and firearm homicide attempts are approximately 20% fatal (40), studying only fatal outcomes misses the morbidity of nonfatal assaults and suicide attempts. We were limited to contrasting age-standardized mortality rates, but would have rather contrasted age-specific rates, had there been more deaths. We did not examine women because their risks of homicide and suicide are much lower than for men, and data for black women were suppressed for more than half of the states. We did not examine Hispanics because they are subject to higher rates of misclassification on death certificate than blacks and whites (41), and extracting Hispanics from the black/white grouping would have cut the data too sparsely. Finally, these results are purely descriptive and make no claims regarding causal mechanisms. 

*Future Research Directions*

This work elicits questions for future investigation. In particular, we would like to better understand the wide variation in black homicide rates, especially across Southern and rust-belt states with similar levels of gun ownership. Multi-level models that take into account state, county, and individual-level characteristics could also shed light on the reasons for the variation in homicide and suicide across states.

*Clinical Implications*

Firearms feature differently in the deaths of black and white men in the United States. The large variation in the homicide and suicide rates and their racial differences across states places a spotlight on areas where policies may help reduce homicide and suicide rates, and their racial inequalities. 

##Figure Legends 
###Figure 1: Comparison of firearm and non-firearm homicide rates between black and white non-Hispanic men by US state, 2008-2016

```{r figure1, fig.width=6, fig.height=8.5, echo = F, warning=F, eval=T}
figure.1 <- ggplot(CDC_Males_AllAges_0816_wide, aes(y = ST.order.RD.homicide, x = adjusted.homicide.White)) + 
  facet_wrap( ~ type) +
  geom_segment(aes(x = adjusted.homicide.White, xend = adjusted.homicide.Black, #the purple line segment between the black and white rate
                   y = ST.order.RD.homicide, yend = ST.order.RD.homicide), 
               lwd = 0.5, col = "#8A5C7B", alpha = 0.5) +
  geom_point(aes(fill = "NH White", size = average.pop.white), pch = 21, col = "black") + #the white homicide rate. For this point type, colour is the *outline*
  geom_point(aes(x = adjusted.homicide.Black, fill = "NH Black", size = average.pop.black), pch = 21, col = "black") + #the black rate
  geom_point(data = CDC_Males_AllAges_0816_wide %>% filter(ST == "US"),   #the US data for whites, shown with a black vertical line
            pch = 73, col = "black", size = 4) +
  geom_point(data =CDC_Males_AllAges_0816_wide %>% filter(ST == "US"), #the US data for blacks, shown with a purple vertical line
             aes(x = adjusted.homicide.Black), 
             pch = 73, col = "#8A5C7B", size = 4) + 
  scale_fill_manual(values = c("#8A5C7B", "white"), labels = c("Black", "White")) + #manually set the *fill* colours for the circles to purple and white
  scale_alpha_continuous(guide = F) +
  scale_size_area(breaks = c(10000, 100000, 500000, 1500000, 3000000, 7000000)) + #define what areas to show in the legend
  ylab("") +
  xlab("Homicide rate (per 100,000 per year)") + 
  theme_bw() + 
  guides(fill = guide_legend(title = "Race"), size = guide_legend(title = "Race-specific\npopulation")) + 
  theme(legend.position = c(0.87, 0.20), 
        plot.title = element_text(size = 10),
        legend.background = element_rect(fill=alpha('white', 0.2)),
        strip.background = element_rect(fill="white"), 
        strip.text.x = element_text(size = 12),
        strip.text.y = element_text(size = 12)) + 
   geom_text(data =CDC_Males_AllAges_0816_wide %>% filter(is.na(adj.RD.homicide) == F), #add labels
             aes(x = adjusted.homicide.Black + 8, label = paste0("+", round(adj.RD.homicide))),
             size = 3, col = "#754255") +
  scale_x_continuous(limits = c(0, 70))
```

The homicide rate for non-Hispanic white men (white circle) can be compared to the homicide rate for non-Hispanic black men (purple circle) for each state. The area of the circle is proportional to its race-specific population size, with the size legend depicting sizes for a range of population sizes. For each state, the difference between the rates is listed to the right of the black homicide rate and quantifies the excess number of homicides to black men per 100,000 per year. The average firearm and non-firearm homicide rates for the entire US are also shown at the top of each graph, with the remaining states organized according to the size of the black-white difference. States that are missing circles for black and/or white men had 19 or fewer firearm or non-firearm homicides during the time period within the affected racial group and were not available for analysis. 

###Figure 2: Comparison of firearm and non-firearm suicide rates between black and white non-Hispanic men by US state, 2008-2016

```{r figure2, fig.width=6, fig.height=8.5, echo = F, warning=F, eval=T}
figure.2 <- ggplot(CDC_Males_AllAges_0816_wide, aes(y = ST.order.RD.suicide, x = adjusted.suicide.White)) + 
  geom_segment(aes(x = adjusted.suicide.White, xend = adjusted.suicide.Black, 
                   y = ST.order.RD.suicide, yend = ST.order.RD.suicide), 
               lwd = 0.5, alpha = 0.5) +
  geom_point(aes(fill = "NH White", size = average.pop.white), pch = 21) + 
  geom_point(aes(x = adjusted.suicide.Black, fill = "NH Black", size = average.pop.black), pch = 21) + 
  geom_point(data =CDC_Males_AllAges_0816_wide %>% filter(ST == "US"),  
            pch = 73, col = "black", size = 4) +
  geom_point(data =CDC_Males_AllAges_0816_wide %>% filter(ST == "US"), 
             aes(x = adjusted.suicide.Black), 
             pch = 73, col = "#8A5C7B", size = 4) + 
  scale_fill_manual(values = c("#8A5C7B", "white"), labels = c("Black", "White")) +
  scale_alpha_continuous(guide = F) +
  scale_size_area(breaks = c(10000, 100000, 500000, 1500000, 3000000, 7000000)) +
  ylab("") + xlab("Suicide rate (per 100,000 per year)") + theme_bw() + 
  guides(fill = guide_legend(title = "Race"), size = guide_legend(title = "Race-specific\npopulation")) + 
  theme(legend.position = c(0.90, 0.22), 
        plot.title = element_text(size = 10),
        legend.background = element_rect(fill=alpha('white', 0.2)),
        strip.background = element_rect(fill="white"), 
        strip.text.x = element_text(size = 12),
        strip.text.y = element_text(size = 12)) + 
  scale_x_continuous(limits = c(0, 35)) +
  geom_text(data = subset(CDC_Males_AllAges_0816_wide, is.na(adj.RD.suicide)==F), 
            aes(x = adjusted.suicide.White + 4, label = ifelse(!(ST == "DC" & type == "Firearm"), paste0("+", round(-adj.RD.suicide)), round(-adj.RD.suicide))), 
            size = 3) +
  facet_wrap(~type)
```

The suicide rate for non-Hispanic white men (white circle) can be compared to the suicide rate for non-Hispanic black men (purple circle) for each state. The area of the circle is proportional to its race-specific population size, with the size legend depicting sizes for a range of population sizes. For each state, the difference between the rates is listed to the right of the white homicide rate and quantifies the excess number of suicides to white men per 100,000 per year. The average firearm and non-firearm suicide rates for the entire US are also shown at the top of each graph, with the remaining states organized according to the size of the white-black difference. States that are missing circles for black and/or white men had 19 or fewer firearm or non-firearm suicides during the time period within the affected racial group and were not available for analysis.

###Figure 3: The relationship between the annual rates of firearm homicide and suicide among white men by state and reported household firearm ownership, 2008-2016

```{r Figure3, fig.width=8, fig.height=6, echo = F, warning=F, eval = T}
#note: removed DC from the subset because we dont have gun ownership data for DC and including it makes a NA in the fill legend.

figure.3 <- ggplot(subset(CDC_Males_AllAges_0816_wide, type == "Firearm" & !(ST %in% c("US", "DC"))), aes(adjusted.suicide.White, adjusted.homicide.White)) + 
  geom_point(aes(fill = gun.house.prev.category, size = average.pop.white), col = "black", pch = 21, alpha = 0.9) + #add the points sized by population
  geom_text_repel(aes(label = ST), size = 4, min.segment.length = 1) + #add the labels _repel limits overlap
  scale_fill_manual(values = c("#fef0d9", "#fdcc8a", "#fc8d59", "#d7301f")) + #orange colour ramp
  ylab("White homicide rate (per 100,000 per year)") +
  xlab("White suicide rate (per 100,000 per year)") + 
  scale_y_continuous(limits = c(0, 5)) +
  scale_x_continuous(limits = c(0, 30)) +
  theme_bw() +
  theme() + 
  scale_size_area(breaks = c(200000, 500000, 1500000, 3000000, 7000000), 
                  labels = c("200k", "500k", "1.5m", "3m", "7m"),
                  max_size = 15) +
  guides(fill = guide_legend(title = "Gun ownership", 
                             override.aes = list(size = 5), order = 1), 
         size = guide_legend(title = "White population"), order = 2) + 
  geom_text(aes(x = 25, y = 0.2, 
                label = paste0("Spearman's correlation: ", round(spearman.white$spearmans, 2))), 
            check_overlap = T)
```

Each state's homicide and suicide mortality rates among non-Hispanic white men are plotted. The area of the circle is proportional to the white men's population size and the color represents the state gun ownership prevalence. States with suppressed estimates for white men for either firearm homicide or firearm suicide are not plotted. These states are the District of Columbia and Hawaii.

###Figure 4: The relationship between the annual rates of firearm homicide and suicide among black men by state and reported household firearm ownership, 2008-2016

```{r figure4, fig.width = 6, fig.height = 6, echo=F, warning=F, eval = T}
figure.4 <- ggplot(subset(CDC_Males_AllAges_0816_wide, type == "Firearm" & !(ST %in% c("US", "DC"))), aes(x = adjusted.suicide.Black, y = adjusted.homicide.Black)) + 
  geom_point(aes(fill = gun.house.prev.category, size = average.pop.black), col = "black", pch = 21, alpha = 0.9) +
  geom_text_repel(aes(label = ST), size = 4, check_overlap = T) +
  scale_fill_manual(values = c("#fef0d9", "#fdcc8a", "#fc8d59", "#d7301f")) +
  ylab("Black homicide rate (per 100,000 per year)") +
  xlab("Black suicide rate (per 100,000 per year)") +
  scale_y_continuous(limits = c(0, 60)) +
  scale_x_continuous(limits = c(0, 12)) +
  theme_bw() +
  #theme(legend.position = "bottom") +
  scale_size_area(breaks = c(20000, 100000, 500000, 1400000), 
                  labels = c("20k", "100k", "500k", "1.4m"), 
                  max_size = 15) +
  guides(fill = guide_legend(title = "Gun ownership", 
                             override.aes = list(size = 5), order = 1), 
         size = guide_legend(title = "Black population"), order = 2) + 
  geom_text(aes(x = 10, y = 2, label = paste0("Spearman's correlation: ", round(spearman.black$spearmans, 2))), check_overlap = T)
```

```{r save-figures, echo=F}
ggsave(filename = "../Plots/figure1.jpeg", plot = figure.1, device = "jpeg", width = 6, height = 9, units = "in")
ggsave(filename = "../Plots/figure2.jpeg", plot = figure.2, device = "jpeg", width = 6, height = 9, units = "in")
ggsave(filename = "../Plots/figure3.jpeg", plot = figure.3, device = "jpeg", width = 8, height = 6, units = "in")
ggsave(filename = "../Plots/figure4.jpeg", plot = figure.4, device = "jpeg", width = 8, height = 6, units = "in")
```

Each state's homicide and suicide mortality rates among non-Hispanic black men are plotted. The area of the circle is proportional to the black men's population size and the color represents the state gun ownership prevalence. States with suppressed estimates for black men for either firearm homicide or firearm suicide are not plotted. These states are Hawaii, Idaho, Iowa, Maine, Montana, North Dakota, New Hampshire, South Dakota, Rhode Island, Utah, Vermont, and Wyoming.

**Acknowledgements:** Corinne Riddell is a post-doctoral researcher at McGill University, and received salary support from the Department of Epidemiology, Biostatistics, and Occupational Health. Sam Harper is supported by a Chercheur Boursier Junior 2 from the Fonds de la Recherche en Santé du Québec. Magdalena Cerdá is a CAMPOS Faculty Scholar, ADVANCE Program at the University of California, Davis and supported by the National Institutes of Health (#1R21 DA041154-01).

###References

1. Centers for Disease Control and Prevention, National Center for Health Statistics. Underlying Cause of Death 1999-2016 on CDC WONDER Online Database, released December, 2017. Data are from the Multiple Cause of Death Files, 1999-2016, as compiled from data provided by the 57 vital statistics jurisdictions through the Vital Statistics Cooperative Program. Available from: http://wonder.cdc.gov/ucd-icd10.html. [cited 17 Jan 2018]

2. Williams DR, Jackson PB. Social sources of racial disparities in health. *Health Affairs*. 2005 Mar 1;24(2):325-34.

3. Miller M, Azreal D, Hemenway D. Firearms and violent death in the United States. In: Webster DW, Vernick JS, eds. *Reducing Gun Violence in America. Informing Policy with Evidence and Analysis*. Baltimore: The Johns Hopkins University Press; 2013:14-15.

4. Miller M, Azrael D, Hemenway D. Rates of household firearm ownership and homicide across us regions and states, 1988–1997. *American Journal of Public Health*. 2002;92(12):1988-1993.

5. Miller M, Hemenway D, Azrael D. State-level homicide victimization rates in the US in relation to survey measures of household firearm ownership, 2001–2003. *Social Science & Medicine*. 2007;64(3):656-664.

6. Siegel M, Ross CS, King III C. The relationship between gun ownership and firearm homicide rates in the united states, 1981–2010. *American Journal of Public Health*. 2013.

7. Miller M, Barber C, White RA, Azrael D. Firearms and suicide in the United States: Is risk independent of underlying suicidal behavior? *American Journal of Epidemiology*. 2013;178(6):946-955.

8. Miller M, Lippmann SJ, Azrael D, Hemenway D. Household firearm ownership and rates of suicide across the 50 United States. *Journal of Trauma and Acute Care Surgery*. 2007;62(4):1029-1035.

9. Kellermann AL, Rivara FP, Rushforth NB, et al. Gun ownership as a risk factor for homicide in the home. *New England Journal of Medicine*. 1993;329(15):1084-1091.

10. Kellermann AL, Rivara FP, Somes G, et al. Suicide in the home in relation to gun ownership. *New England Journal of Medicine*. 1992;327(7):467-472.

11. Kubrin CE, Wadsworth T. Explaining suicide among blacks and whites: How socioeconomic factors and gun availability affect race-specific suicide rates. *Social Science Quarterly*. 2009;90(5):1203-1227.

12. Loeber R, Farrington DP. Young Male Homicide Offenders and Victims: Current Knowledge, Beliefs, and Key Questions. In: Young Homicide Offenders and Victims. Springer; 2011. p. 1-17.

13. Gibbs JT. African-American suicide: A cultural paradox. *Suicide and life-threatening behavior*. 1997. 27(1):68–79.

14. Centers for Disease Control and Prevention, National Center for Health Statistics. Underlying Cause of Death 1999-2016 on CDC WONDER Online Database, released December, 2017. Data are from the Multiple Cause of Death Files, 1999-2016, as compiled from data provided by the 57 vital statistics jurisdictions through the Vital Statistics Cooperative Program. Accessed at http://wonder.cdc.gov/ucd-icd10.html on Jan 16, 2018.

15. Monuteaux MC, Lee LK, Hemenway D, Mannix R, Fleegler EW. Firearm ownership and violent crime in the us: An ecologic study. *American journal of Preventive Medicine*. 2015;49(2):207-214.

16. R Core Team. *R: A Language and Environment for Statistical Computing*. Vienna, Austria: R Foundation for Statistical Computing; 2014. http://www.R-project.org/.

17. Canadian Institutes of Health Research, National Sciences and Engineering Research Council of Canada, and Sciences Sciences and Humanities Research Council of Canada. *Tri-council policy statement: Ethical conduct for research involving humans*, December 2014.

18. Branas CC, Nance ML, Elliott MR, Richmond TS, Schwab CW. Urban–rural shifts in intentional firearm death: Different causes, same results. *American Journal of Public Health*. 2004 Oct;94(10):1750-5.

19. Cubbin C, Pickle LW, Fingerhut L. Social context and geographic patterns of homicide among US black and white males. American Journal of Public Health. 2000 Apr;90(4):579.

20. Hirsch JK. A review of the literature on rural suicide: Risk and protective factors, incidence, and prevention. *Crisis*. 2006;27(4):189.

21. Wilson WJ. *The Truly Disadvantaged: The Inner City, the Underclass, and Public Policy*. University of Chicago Press; 1987.

22. Massey D. *Problem of the Century*. New York: Russell Sage; 2001:314-344.

23. Ousey GC, Kubrin CE. Exploring the connection between immigration and violent crime rates in US cities, 1980–2000. *Social Problems*. 2009;56(3):447-473.

24. Shihadeh ES, Flynn N. Segregation and crime: The effect of black social isolation on the rates of black urban violence. *Social Forces*. 1996;74(4):1325-1352.

25. Peterson RD, Krivo LJ. Racial segregation and black urban homicide. *Social Forces*. 1993;71(4):1001-1026.

26. Blumstein A. Youth violence, guns, and the illicit-drug industry. *The Journal of Criminal Law and Criminology*. 1995;86(1):10-36.

27. Fryer RG, Heaton PS, Levitt SD, Murphy KM. Measuring crack cocaine and its impact. *Economic Inquiry*. 2013;51(3):1651-1681.

28. Fleegler EW, Lee LK, Monuteaux MC, Hemenway D, Mannix R. Firearm legislation and firearm fatalities in the United States. *JAMA Internal Medicine*. 2013;173(9):732-40.

29. Crifasi CK, Meyers JS, Vernick JS, Webster DW. Effects of changes in permit-to-purchase handgun laws in Connecticut and Missouri on suicide rates. *Preventive Medicine*. 2015;79:43-9.

30. Rudolph KE, Stuart EA, Vernick JS, Webster DW. Association between Connecticut’s permit-to-purchase handgun law and homicides. *American Journal of Public Health*. 2015;105(8):e49-54.

31. Spicer RS, Miller TR. Suicide acts in 8 states: Incidence and case fatality rates by demographics and method. *American Journal of Public Health*. 2000;90(12):1885.

32. Miller M, Azrael D, Hemenway D. The epidemiology of case fatality rates for suicide in the Northeast. *Annals of Emergency Medicine*. 2004;43(6):723-730.

33. Siegel M, Pahn M, Xuan Z, et al. firearm laws in all 50 US states, 1991–2016. *American Journal of Public Health*. 2017;(0):e1-e8.

34. Mościcki EK. Identification of suicide risk factors using epidemiologic studies. *Psychiatric Clinics of North America*. 1997;20(3):499-517.

35. Mościcki EK. Epidemiology of completed and attempted suicide: Toward a framework for prevention. *Clinical Neuroscience Research*. 2001;1(5):310-323.

36. Centers for Disease Control and Prevention. Alcohol and suicide among racial/ethnic populations-17 states, 2005-2006. *MMWR: Morbidity and Mortality Weekly Report*. 2009;58(23):637-41.

37. Wintemute GJ. The epidemiology of firearm violence in the twenty-first century United States. *Annual Review of Public Health*. 2015;18(36):5-19.

38. Centers for Disease Control and Prevention. Homicides --- United States, 2007 and 2009. *MMWR: Morbidity and Mortality Weekly Report*. 2013;62(Suppl 3):164-170.

39. Rossen LM, Bastian B, Warner M, Khan D, Chong Y. Drug poisoning mortality: United States, 1999–2016. National Center for Health Statistics. 2017. Accessed at:  https://www.cdc.gov/nchs/data-visualization/drug-poisoning-mortality/) on January 15, 2018.

40. Cook PJ, Rivera-Aguirre AE, Cerdá M, Wintemute G. Constant lethality of gunshot injuries from firearm assault: United states, 2003–2012. *American Journal of Public Health*. 2017;107(8):1324-1328.

41. Arias E, Schauman WS, Eschbach K, Sorlie PD, Backlund E. The validity of race and Hispanic origin reporting on death certificates in the United States. National Center for Health Statistics. *Vital and Health Statistics*. 2008;2(148):1-23.
